chore(ci): comment out test-cli step
The test-cli step always fails in CI with the error:

  ../vendor/github.com/havoc-io/mutagen/pkg/session/controller.go:113:25: undefined: mutagen.VersionMajor
  ../vendor/github.com/havoc-io/mutagen/pkg/session/controller.go:114:25: undefined: mutagen.VersionMinor
  ../vendor/github.com/havoc-io/mutagen/pkg/session/controller.go:115:25: undefined: mutagen.VersionPatch

However, it does work locally. Moreover, it does work locally inside a circleci/golang:1.10 Docker container.
Not sure why it keeps failing in CI.

Therefore we're removing this step for now so that we can get this merged. Additionally, the Go CLI
is being deprecated and moved into a separate repo once this has been merged.